% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Logistic regression}
\usage{
logistic_regression(X, y, maxi = 100, threshold = 1e-10)
}
\arguments{
\item{X}{input predictors}

\item{y}{input binary responses taking value 0 and 1}

\item{maxi}{the maximum of iteration}

\item{threshold}{the tolerance of precision}
}
\value{
the estimated parameters(beta)
}
\description{
Find the estimated parameters of the logistic regression model.
}
\examples{
X <- cbind(rbinom(30,1,0.4),rnorm(30,3,1))
y <- rbinom(30, 1, 0.45)
logistic_regression(X,y)

}
